add RGBDS's HIGH and LOW functions
HIGH() and LOW() return high and low byte of a word (2-bytes) or
returns the High / Low register from a register pair, respectively. So
for example, HIGH(BC) will be replaced with B. LOW(BC) will be replaced
with C. HIGH($FF89) will replace with $FF and LOW($FF89) will replace
with $89. This can be used inside a macro to split apart an argument
into its high and low bytes / register. Very handy.